@dhardy 2024-08-12
This appears to be an alternative to #1484.