@benjamin-lieser 2024-08-15
The `n==1` case is trivial, no binv needed:
```
let u : f64 = rng.random();
if u < p {1} else {0}
```

I also think we can unify the code paths.
Also not sure if this catches the `q == 1.0` problem